refactor: merge squeeze partial encoder/decoder into `SqueezeCodecPartial`
